Hook up Allow Mock Emitter toggle button.
