+++ UnityExamples/Assets/Examples/ExamplesAssets/SensationExplorer/Scripts/AllowMockEmitterToggle.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 2594efd561c63724384b08b5fb3c25d4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
